[S] Refactor following the new code structure
